Stop threading the visibility toggle through the shared VertexStyleRow
VertexStyleRow is also rendered by the main Graph view's Styles panel,
so a showVisibilityToggle prop and an unconditional useHiddenSchemaTypes()
subscription inside it coupled the shared row to a Schema-view-only
feature. Wrap it from the outside instead (SchemaVertexStyleRow in
VertexStyles.tsx), leaving VertexStyleRow itself untouched.

Also tighten two new tests to assert full expected values with
toStrictEqual instead of toContain/toEqual, per docs/agents/testing.md.
