+++ packages/graph-explorer/src/modules/Styles/VertexStyleRow.tsx
@@ -1,33 +1,19 @@
-import { EyeIcon, EyeOffIcon } from "lucide-react";
 import { type ComponentPropsWithRef, useEffect, useState } from "react";
 
 import { Button, FormItem, InputField, Label, StylingIcon } from "@/components";
-import {
-  useDisplayVertexTypeConfig,
-  useHiddenSchemaTypes,
-  type VertexType,
-} from "@/core";
+import { useDisplayVertexTypeConfig, type VertexType } from "@/core";
 import { useVertexStyling } from "@/core/StateProvider/graphStyles";
 import { useDebounceValue, usePrevious } from "@/hooks";
 import { useOpenNodeStyleDialog } from "@/modules/NodesStyling";
-import { cn } from "@/utils";
 import { LABELS } from "@/utils/constants";
 
 export type VertexStyleRowProps = {
   vertexType: VertexType;
-  /** Shows an eye toggle to hide this vertex type from the Schema view. */
-  showVisibilityToggle?: boolean;
 } & ComponentPropsWithRef<typeof FormItem>;
 
-export function VertexStyleRow({
-  vertexType,
-  showVisibilityToggle,
-  className,
-  ...rest
-}: VertexStyleRowProps) {
+export function VertexStyleRow({ vertexType, ...rest }: VertexStyleRowProps) {
   const { setVertexStyle } = useVertexStyling(vertexType);
   const displayConfig = useDisplayVertexTypeConfig(vertexType);
-  const { isHidden, toggleType } = useHiddenSchemaTypes();
 
   const [displayAs, setDisplayAs] = useState(displayConfig.displayLabel);
 
@@ -44,10 +30,8 @@ export function VertexStyleRow({
     setVertexStyle({ displayLabel: debouncedDisplayAs });
   }, [debouncedDisplayAs, prevDisplayAs, setVertexStyle]);
 
-  const hidden = showVisibilityToggle && isHidden(vertexType);
-
   return (
-    <FormItem className={cn(hidden && "opacity-50", className)} {...rest}>
+    <FormItem {...rest}>
       {vertexType ? (
         <Label>{vertexType}</Label>
       ) : (
@@ -66,20 +50,6 @@ export function VertexStyleRow({
           <StylingIcon />
           Customize
         </Button>
-        {showVisibilityToggle ? (
-          <Button
-            variant="ghost"
-            size="icon-small"
-            tooltip={
-              hidden
-                ? `Show ${vertexType} in schema view`
-                : `Hide ${vertexType} from schema view`
-            }
-            onClick={() => toggleType(vertexType)}
-          >
-            {hidden ? <EyeOffIcon /> : <EyeIcon />}
-          </Button>
-        ) : null}
       </div>
     </FormItem>
   );
